Extract findMethodsToOverride as an utility
So that when and mock can share it
